moved requestType classes
